+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to
   * `isAutoNegotiationSettable`
   * `isPHYResettable`
 * Added call to `qnethernet_hal_deinit_entropy()` in `RandomDevice` destructor.
+* Added `RandomDevice::entropy()` to match `std::random_device::entropy()`.
 
 ### Changed
 * Improved _Ping_ example by always sending, even if a send fails. It was